Move Document publish action to the server
See #40.
