+++ lib/document.js
@@ -61,32 +61,6 @@ Document.Meta({
   },
 });
 
-Document.publish = method(new ValidatedMethod({
-  name: "Document.publish",
-
-  validate(args) {
-    check(args, {
-      documentId: String,
-    });
-  },
-
-  run(args) {
-    const user = Meteor.user(User.REFERENCE_FIELDS());
-    if (!user) {
-      throw new Meteor.Error('unauthorized', "Unauthorized.");
-    }
-
-    Document.documents.update({
-      _id: args.documentId,
-    }, {
-      $set: {
-        publishedBy: user.getReference(),
-        publishedAt: new Date(),
-      },
-    });
-  },
-}));
-
 Document.create = method(new ValidatedMethod({
   name: 'Document.create',
 
@@ -122,6 +96,11 @@ Document.create = method(new ValidatedMethod({
     };
   },
 }));
+
+Document.publish = function publishDocument(...args) {
+  args.unshift('Document.publish');
+  return Meteor.call(...args);
+};
 
 if (Meteor.isServer) {
   Document.Meta.collection._ensureIndex({
