+++ translations/locale/en_US/LC_MESSAGES/app.po
@@ -7,7 +7,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: peer-doc undefined\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2018-03-19 16:24-0300\n"
+"POT-Creation-Date: 2018-03-19 16:57-0300\n"
 "PO-Revision-Date: 2018-03-05 12:12-0300\n"
 "Last-Translator: Automatically generated\n"
 "Language-Team: none\n"
@@ -109,7 +109,11 @@ msgstr "Enter a link"
 msgid "invalid-url"
 msgstr "Invalid URL."
 
-#: client/components/publish-confirmation.vue:53
+#: client/components/publish-confirmation.vue:67
+msgid "publish-error"
+msgstr "An error has occurred when trying to publish the document."
+
+#: client/components/publish-confirmation.vue:72
 msgid "publish-success"
 msgstr "The document was successfully published."
