+++ translations/translations.json
@@ -1,1 +1,1 @@
-{"en_US":{"access-denied":"Access denied.","cancel":"Cancel","chat":"Chat","close":"Close","comments":"Comments","documents":"Documents","draft":"Draft","front-page":"Front page.","history":"History","insert":"Insert","no-documents":"No documents.","not-found":"Not found.","remove":"Remove","sign-in":"Sign In","sign-out":"Sign Out","untitled":"Untitled","username-is-required":"Username is required","username-too-short":"Username too short, it should be 4 characters or more","username-invalid":"Invalid username, it should contain only basic characters","username":"Username","signed-in-success":"You have been signed in.","link-hint":"Enter a link","invalid-url":"Invalid URL.","publish-success":"The document was successfully published.","document-created-error":"Error creating a new document: %{error}","document-created-success":"New document has been created.","signed-out-error":"Error signing out: %{error}","signed-out-success":"You have been signed out."}}
+{"en_US":{"access-denied":"Access denied.","cancel":"Cancel","chat":"Chat","close":"Close","comments":"Comments","documents":"Documents","draft":"Draft","front-page":"Front page.","history":"History","insert":"Insert","no-documents":"No documents.","not-found":"Not found.","remove":"Remove","sign-in":"Sign In","sign-out":"Sign Out","untitled":"Untitled","username-is-required":"Username is required","username-too-short":"Username too short, it should be 4 characters or more","username-invalid":"Invalid username, it should contain only basic characters","username":"Username","signed-in-success":"You have been signed in.","link-hint":"Enter a link","invalid-url":"Invalid URL.","publish-error":"An error has occurred when trying to publish the document.","publish-success":"The document was successfully published.","document-created-error":"Error creating a new document: %{error}","document-created-success":"New document has been created.","signed-out-error":"Error signing out: %{error}","signed-out-success":"You have been signed out."}}
